+++ emulator/emulator.go
@@ -6,11 +6,11 @@ import (
 	"math/rand"
 	"time"
 
-	"github.com/bchadwic/chip8/emulator/display"
-	"github.com/bchadwic/chip8/emulator/display/emit"
-	"github.com/bchadwic/chip8/emulator/drivers"
-	"github.com/bchadwic/chip8/emulator/keypad"
-	"github.com/bchadwic/chip8/emulator/speaker"
+	"github.com/bchadwic/chip8/internal/display"
+	"github.com/bchadwic/chip8/internal/display/emit"
+	"github.com/bchadwic/chip8/internal/drivers"
+	"github.com/bchadwic/chip8/internal/keypad"
+	"github.com/bchadwic/chip8/internal/speaker"
 )
 
 const (
