+++ internal/display/emit/emit.go
