+++ internal/speaker/speaker_test.go
